# Conflicts:
#	WEB/web-admin/src/api/teacher.js
#	WEB/web-admin/src/config/index.js

+++ JAVA/onlineExamination/src/main/java/com/wxj/model/VO/ExamQuestionsDetailsVO.java
@@ -14,7 +14,7 @@ public class ExamQuestionsDetailsVO implements Serializable {
     /**
      * 题号
      */
-    private Integer questionsNo;
+    private Byte questionsNo;
     /**
      * ID
      */
